add openrc support
