+++ sources/boot_service/__init__.py
@@ -1,0 +1,1 @@
+from boot_service.base_boot_service import BaseBootService
