+++ sources/boot_service/openrc/__init__.py
@@ -1,0 +1,2 @@
+from boot_service.openrc.openrc import Openrc
+openrc = Openrc
