+++ sources/boot_service/openrc/linux-enable-ir-emitter
