+++ sources/boot_service/systemd/__init__.py
@@ -1,0 +1,2 @@
+from boot_service.systemd.systemd import Systemd
+systemd = Systemd
