+++ sources/command/boot.py
@@ -1,15 +1,11 @@
 import logging
 from typing import NoReturn
 
-from globals import ExitCode, get_devices
-from openrc import Openrc
-from systemd import Systemd
-
-from ..service.service_manager import get_service_manager
+from globals import ExitCode, get_devices, get_boot_service_constructor
 
 
 def boot(boot_status: str) -> NoReturn:
-    """Enable or disable the systemd service which
+    """Enable or disable the boot service which
     activates the ir emitter for all configured device
 
     args:
@@ -21,26 +17,29 @@ def boot(boot_status: str) -> NoReturn:
     """
     assert boot_status in ["enable", "disable", "status"]
 
-    devices = get_devices()
-    service_manager = get_service_manager(devices)
-    if service_manager is None:
+    boot_service_constructor = get_boot_service_constructor()
+    if boot_service_constructor is None:
+        logging.critical("No supported boot service are installed.")
         exit(ExitCode.FAILURE)
+
+    devices = get_devices()
+    boot_service = boot_service_constructor(devices)
 
     if boot_status == "enable":
         if not len(devices):
             logging.critical("No driver have been configured.")
             exit(ExitCode.FAILURE)
-        if service_manager.enable():
+        if boot_service.enable():
             exit(ExitCode.FAILURE)
         logging.info("The boot service have been enabled.")
 
     elif boot_status == "disable":
-        if service_manager.disable():
+        if boot_service.disable():
             exit(ExitCode.FAILURE)
         logging.info("The boot service have been disabled.")
 
     elif boot_status == "status":
-        if service_manager.status():
+        if boot_service.status():
             exit(ExitCode.FAILURE)
 
     exit(ExitCode.SUCCESS)
