--- sources/service/systemd/__init__.py
@@ -1,1 +1,0 @@
-from systemd.Systemd import Systemd
